refactor(io): move io-only helpers into io utils
